+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [#690](https://github.com/estruyf/vscode-front-matter/issues/690): Added the ability to filter values in the `contentRelationship` field
 - [#700](https://github.com/estruyf/vscode-front-matter/issues/700): Added the `{{pathToken.relPath}}` placeholder for the `previewPath` property
 - [#706](https://github.com/estruyf/vscode-front-matter/issues/706): Show the error of scripts failing in the Front Matter output panel
+- [#709](https://github.com/estruyf/vscode-front-matter/issues/709): Take "where clause" into account on content creation
 
 ### ⚡️ Optimizations